% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drawLogit.R
\name{drawLogit}
\alias{drawLogit}
\title{drawLogit}
\usage{
drawLogit(effectSize = 1, n = 100, p = 4, indicesTrulyPredictive = 1)
}
\arguments{
\item{effectSize}{the mean difference in the truly predictive genes. Default is 0.7.}

\item{n}{sample size}

\item{p}{Number of features}

\item{indicesTrulyPredictive}{the indices of the truly predictive genes.}
}
\value{
a dataframe with the following variables:
itemize{
\item{"x"}{the design matrix containing the values of the covariables. Independent observations are in the rows, covariables in the columns.}
\item{"y"}{the response vector containing information about the binary indepedent variable.}
}
}
\description{
drawsDataFromLogitModel
}
\examples{
drawLogit(n=5, p=2)
}

