% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simTestSize.R
\name{simTestSize}
\alias{simTestSize}
\title{simTestSize: main function of package simTestSize}
\usage{
simTestSize(learningAlgorithm, drawFunction, g, conf.level = 0.95,
  testChunkSize, Nwithin = 110, Nbetween = 100, ...)
}
\arguments{
\item{learningAlgorithm}{a function that takes learning data as argument and outputs a prediction rule in the form of another function that maps test data to numeric vectors. These can be errors or AUCs, one for each testing data set.}

\item{drawFunction}{a function that returns a dataframe with the response variable in a column names y}

\item{g}{the learning sample sizelearningAlgorithm, drawFunction, Nwithin=12, Nbetween=17)}

\item{conf.level}{the confidence level; defaults to .95}

\item{testChunkSize}{the sample size of a single test chunk}

\item{Nwithin}{number of test chunks to be drawn for each training iteration. In the paper, this number was called n_test.}

\item{Nbetween}{number of training data sets to be drawn. In the paper, this number was calles N.}

\item{...}{additional parameters to be passed to the drawFunction}
}
\value{
a list with entries lower, point, upper for the confidence interval for the ratio of the two variances and another for the optimal number of test chunks, the error matrix, the times C and B, and the variance within learning sets (vwl), and the variance between learning sets (vbl).
}
\description{
computes a suggested best number of test chunks, i.e., calls to the function passed to the argument drawFunction.
}
\examples{
simTestSize(
learningAlgorithm=svmLearning,
drawFunction=drawLogit,
g=20,
testChunkSize=1
)
}

