% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stopwatch.R
\name{stopwatch}
\alias{stopwatch}
\title{stopwatch}
\usage{
stopwatch(learningAlgorithm, drawFunction, g, testChunkSize, ...)
}
\arguments{
\item{learningAlgorithm}{a function that takes learning data as argument and outputs a prediction rule in the form of another function that maps test data to numeric vectors. These can be errors or AUCs, one for each testing data set.}

\item{drawFunction}{a function that returns a dataframe with the response variable in a column names y}

\item{g}{the learning sample sizelearningAlgorithm, drawFunction, Nwithin=12, Nbetween=17)}

\item{testChunkSize}{the sample size of a single test chunk}

\item{...}{additional parameters to be passed to the drawing function}
}
\value{
a list with the times A and B in nanoseconds
}
\description{
measures computer time for learning and testing.
There is a difficulty with the lazy evaluation mechanism in R: since the learning funciton returns a function that acts as a closure and keeps a reference to its calling environment, and the model is not learnt before actually evaluating the function, we have to measure the times A and B indirectly.
Therefore, this function approximates the learning time by measuring the time for learning and testing k times, in such a way that the evaluation promise has to be fulfilled, minus the time for drawing, for various values of k, and then fitting the linear model that the total time is A + kB.
}
\examples{
stopwatch(
learningAlgorithm=svmLearning,
drawFunction=drawLogit,
g=20,
testChunkSize=1
)
}

