% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/svmLearning.R
\name{svmLearning}
\alias{svmLearning}
\title{svmLearning}
\usage{
svmLearning(trainingData)
}
\arguments{
\item{trainingData}{the training data: a dataframe whose response column is labelled "y"}
}
\value{
a prediction rule in the form of a function that eats testing data and gives predictions
}
\description{
an example learning algorithm: logistic regression
}
\examples{
svmLearning(drawLogit(n=1000, effectSize=5))(drawLogit(n=1000, effectSize=5))
}

